requirements.txt
alembic==1.13.3
amqp==5.2.0
aniso8601==9.0.1
blinker==1.8.2
boto3==1.35.40
botocore==1.35.40
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
firebase-admin==6.5.0
Flask==2.1.2
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==2.5.1
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.20.1
msgpack==1.1.0
numpy==2.3.4
packaging==24.1
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.23.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
shapely==2.1.2
six==1.16.0
SQLAlchemy==1.4.41
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
Werkzeug==2.0.2
